pin the loop variable in table tests
Table tests relies on calling a closure that uses a loop variable where
each for iteration, the variable holds data for a test case. When adding
t.Parallel(), things can and will go south due to a well-known Go
footgun: https://github.com/golang/go/wiki/CommonMistakes#using-goroutines-on-loop-iterator-variables

Oops. This commits adds a function that looks for table-driven tests and
pins the loop variable to fix the problem.